add capability to compute rewards online in during the sampling of batches in the SAC agent training in the train_one_epoch method; this is a lot more efficient for deep rewards and also removes the need to discard old experience once the reward fn gets updated by the IRL routine; this is the case because now we only keep observations and actions during samplig, and compute rewards using the current reward_fn
I was reluctant before since I got small differences when doing this and
precomputing the rewards during the sampling of the episodes. This was
because of floating point precision of +- 1e-6 deviations; I checked it
and indeed in the printouts only these decimal places were different.
There does not seem to be a way to fix the float precision, so I
recommend just sticking to this method because ot the mentioned
benefits.